Creating Folders Ready
Next step is the merge PDF. this I need to find and work out how to use it.

There is a stream that is been printed to the terminal, this is some of the information that I wish to display to the
user at a later date.

Changes to be committed:
	modified:   src/boomatang/CollectFiles.java
	new file:   src/boomatang/CreateFolder.java
	modified:   src/boomatang/Events.java
	new file:   src/boomatang/FolderWalk.java
	modified:   src/boomatang/GUI.java
	modified:   src/boomatang/PdfObj.java
	new file:   src/boomatang/ProcessFiles.java
	modified:   src/boomatang/SortFiles.java